requirements.txt
# PhishGuard AI - Dependencies
flask>=3.0.0
flask-cors>=4.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
shap>=0.44.0
matplotlib>=3.8.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
google-generativeai>=0.5.0
joblib>=1.3.0
python-dotenv>=1.0.0
